# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  version: 1.0.0
  description: Connect with your friends effortlessly using WASAText!
servers:
  - url: http://localhost:8080/v1
...
paths:
  /users:
    get:
      operationId: getUsers
      summary: Get users
      description: Gets a list of users optionally filtered by a search query
      tags:
        - users
      parameters:
        - name: q
          in: query
          required: false
          description: Search query to filter users by username
          schema:
            type: string
            minLength: 1
            maxLength: 16
            pattern: "^.*$"
            description: Search query
          example: Maria
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: List of users object
                properties:
                  users:
                    type: array
                    minItems: 0
                    maxItems: 100
                    description: List of users
                    items:
                      $ref: "#/components/schemas/User"
                required:
                  - users
              example:
                users:
                  - userId: "550e8400-e29b-41d4-a716-446655440000"
                    username: Maria
                    profilePicture: "http://localhost:8080/uploads/profile-pictures/550e8400-e29b-41d4-a716-446655440000.jpg"
                    createdAt: "2023-10-01T12:00:00Z"
                  - userId: "550e8400-e29b-41d4-a716-446655440001"
                    username: John
                    profilePicture: "http://localhost:8080/uploads/profile-pictures/550e8400-e29b-41d4-a716-446655440001.jpg"
                    createdAt: "2023-10-02T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: doLogin
      summary: Login
      description: Login or register a user
      tags:
        - users
      requestBody:
        description: User login information
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username object
              properties:
                username:
                  $ref: "#/components/schemas/Username"
              required:
                - username
            example:
              username: Maria
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                userExample:
                  $ref: "#/components/examples/userExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}:
    get:
      operationId: getUser
      summary: Get user
      description: Gets a user by their unique identifier
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/userId"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                userExample:
                  $ref: "#/components/examples/userExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /me/username:
    put:
      operationId: setMyUserName
      summary: Update username
      description: Update the username of the authenticated user
      tags:
        - users
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username object
              properties:
                username:
                  $ref: "#/components/schemas/Username"
              required:
                - username
            example:
              username: Maria
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                userExample:
                  $ref: "#/components/examples/userExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /me/photo:
    put:
      operationId: setMyPhoto
      summary: Update photo
      description: Update the profile picture of the authenticated user
      tags:
        - users
      requestBody:
        description: User details
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Image object
              properties:
                image:
                  $ref: "#/components/schemas/Image"
              required:
                - image
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile picture updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                userExample:
                  $ref: "#/components/examples/userExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Id:
      type: string
      format: uuid
      minLength: 36
      maxLength: 36
      pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      description: Unique identifier
    Timestamp:
      type: string
      format: date-time
      minLength: 20
      maxLength: 20
      description: Timestamp in RFC 3339 format
    Image:
      type: string
      format: binary
      minLength: 1
      maxLength: 4194304
      description: Binary data for image upload

    # --------------------------------------------------------------------------------
    # User
    Username:
      type: string
      minLength: 3
      maxLength: 16
      pattern: "^[a-zA-Z0-9._]{3,16}$"
      description: Unique handle used to identify the user
    ProfilePicture:
      type: string
      minLength: 11
      maxLength: 255
      pattern: "^http://(localhost|127\\.0\\.0\\.1):[0-9]{1,5}/uploads/profile-pictures/(default|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})\\.(jpg|jpeg|png|gif|webp|bmp|svg)$"
      description: URL of the user's profile picture
    UserPreview:
      type: object
      description: Preview of user details
      properties:
        userId:
          $ref: "#/components/schemas/Id"
        username:
          $ref: "#/components/schemas/Username"
        profilePicture:
          $ref: "#/components/schemas/ProfilePicture"
      required:
        - userId
        - username
    User:
      allOf:
        - $ref: "#/components/schemas/UserPreview"
        - type: object
          description: User details
          properties:
            createdAt:
              $ref: "#/components/schemas/Timestamp"
          required:
            - createdAt

    # --------------------------------------------------------------------------------
    # Conversation
...
    Error:
      type: object
      description: Error details
      properties:
        code:
          type: string
          minLength: 7
          maxLength: 7
          pattern: "^ERR-[0-9]{3}$"
          description: Machine-readable error code
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^.*$"
          description: Brief description of the error
      required:
        - code
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-400"
            message: Bad request
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-401"
            message: Unauthorized access
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-404"
            message: Resource not found
    Conflict:
      description: Conflict error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-409"
            message: Conflict error
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-500"
            message: Internal server error
  examples:
    userExample:
      summary: Example user
      value:
        userId: "550e8400-e29b-41d4-a716-446655440000"
        username: Maria
        profilePicture: "http://localhost:8080/uploads/profile-pictures/550e8400-e29b-41d4-a716-446655440000.jpg"
        createdAt: "2023-10-01T12:00:00Z"
...
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: Unique identifier of the user
      schema:
        $ref: "#/components/schemas/Id"
      example: "550e8400-e29b-41d4-a716-446655440000"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: uuid